+++ NEWS.md
@@ -1,3 +1,12 @@
+## v 0.3.1.9000
+
+---
+
+To be released as 0.3.2.
+
+* Add information about R session timeout option in the error message when
+`utils::download.file()` fails (#13 by `@patrickvossler18`)
+
 ## v 0.3.1
 
 ----